refactor: move s3 module to io
